feat: AttributeConverter 클래스 분리
