+++ src/main/java/kr/pickple/back/member/domain/MemberPosition.java
@@ -12,7 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
 import kr.pickple.back.common.domain.BaseEntity;
 import kr.pickple.back.position.domain.Position;
-import kr.pickple.back.position.domain.Position.PositionConverter;
+import kr.pickple.back.position.util.PositionStatusConverter;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
@@ -26,7 +26,7 @@ public class MemberPosition extends BaseEntity {
     private Long id;
 
     @NotNull
-    @Convert(converter = PositionConverter.class)
+    @Convert(converter = PositionStatusConverter.class)
     @Column(length = 2)
     private Position position;
