+++ src/main/java/kr/pickple/back/member/domain/MemberStatus.java
@@ -11,8 +11,6 @@ import java.util.stream.Stream;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
 import kr.pickple.back.member.exception.MemberException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -38,19 +36,5 @@ public enum MemberStatus {
         }
 
         throw new MemberException(MEMBER_STATUS_NOT_FOUND, description);
-    }
-
-    @Converter
-    public static final class MemberStatusConverter implements AttributeConverter<MemberStatus, String> {
-
-        @Override
-        public String convertToDatabaseColumn(final MemberStatus memberStatus) {
-            return memberStatus.getDescription();
-        }
-
-        @Override
-        public MemberStatus convertToEntityAttribute(final String description) {
-            return MemberStatus.from(description);
-        }
     }
 }
